adding GetBoneTransformation to Animation and now the given a bone a position in worldspace can be calculated. this also fix the magic numbers on the triggers of the brawlers characters that are attached to the models

+++ Engine/Source/Editor/Editor.cpp
@@ -810,7 +810,7 @@ namespace Editor
 	{
 		if (animationSequencer.showing)
 		{
-			ImVec2 WorkSize(width, height);
+			ImVec2 WorkSize(static_cast<float>(width), static_cast<float>(height));
 			ImVec2 seqPos = ImVec2(WorkSize.x / sequencerAdjustment, WorkSize.y / sequencerAdjustment);
 			ImVec2 seqSize = ImVec2(WorkSize.x * (1.0f - (2.0f / sequencerAdjustment)), WorkSize.y * (1.0f - (2.0f / sequencerAdjustment)));
 			animationSequencer.Resize(seqPos, seqSize);
